LayoutTests/imported/w3c/web-platform-tests/fetch/api/request/request-idl-expected.txt


FAIL Request interface: existence and properties of interface object assert_equals: class string of Request expected "[object Function]" but got "[object RequestConstructor]"
PASS Request interface object length 
PASS Request interface object name 
PASS Request interface: existence and properties of interface prototype object 
PASS Request interface: existence and properties of interface prototype object's "constructor" property 
FAIL Request interface: attribute method assert_equals: getter must be Function expected "function" but got "undefined"
FAIL Request interface: attribute url assert_equals: getter must be Function expected "function" but got "undefined"
FAIL Request interface: attribute headers assert_equals: getter must be Function expected "function" but got "undefined"
FAIL Request interface: attribute type assert_equals: getter must be Function expected "function" but got "undefined"
FAIL Request interface: attribute destination assert_equals: getter must be Function expected "function" but got "undefined"
FAIL Request interface: attribute referrer assert_equals: getter must be Function expected "function" but got "undefined"
FAIL Request interface: attribute referrerPolicy assert_equals: getter must be Function expected "function" but got "undefined"
FAIL Request interface: attribute mode assert_equals: getter must be Function expected "function" but got "undefined"
FAIL Request interface: attribute credentials assert_equals: getter must be Function expected "function" but got "undefined"
FAIL Request interface: attribute cache assert_equals: getter must be Function expected "function" but got "undefined"
FAIL Request interface: attribute redirect assert_equals: getter must be Function expected "function" but got "undefined"
FAIL Request interface: attribute integrity assert_equals: getter must be Function expected "function" but got "undefined"
PASS Request interface: operation clone() 
FAIL Request interface: attribute bodyUsed assert_equals: getter must be Function expected "function" but got "undefined"
FAIL Request interface: operation arrayBuffer() assert_throws: calling operation with this = null didn't throw TypeError function "function () {
                memberHolderObject[member.n..." did not throw
FAIL Request interface: operation blob() assert_throws: calling operation with this = null didn't throw TypeError function "function () {
                memberHolderObject[member.n..." did not throw
FAIL Request interface: operation formData() assert_throws: calling operation with this = null didn't throw TypeError function "function () {
                memberHolderObject[member.n..." did not throw
FAIL Request interface: operation json() assert_throws: calling operation with this = null didn't throw TypeError function "function () {
                memberHolderObject[member.n..." did not throw
FAIL Request interface: operation text() assert_throws: calling operation with this = null didn't throw TypeError function "function () {
                memberHolderObject[member.n..." did not throw
PASS Request must be primary interface of new Request("") 
PASS Stringification of new Request("") 
PASS Request interface: new Request("") must inherit property "method" with the proper type (0) 
PASS Request interface: new Request("") must inherit property "url" with the proper type (1) 
PASS Request interface: new Request("") must inherit property "headers" with the proper type (2) 
PASS Request interface: new Request("") must inherit property "type" with the proper type (3) 
PASS Request interface: new Request("") must inherit property "destination" with the proper type (4) 
PASS Request interface: new Request("") must inherit property "referrer" with the proper type (5) 
FAIL Request interface: new Request("") must inherit property "referrerPolicy" with the proper type (6) Unrecognized type ReferrerPolicy
PASS Request interface: new Request("") must inherit property "mode" with the proper type (7) 
PASS Request interface: new Request("") must inherit property "credentials" with the proper type (8) 
PASS Request interface: new Request("") must inherit property "cache" with the proper type (9) 
PASS Request interface: new Request("") must inherit property "redirect" with the proper type (10) 
PASS Request interface: new Request("") must inherit property "integrity" with the proper type (11) 
PASS Request interface: new Request("") must inherit property "clone" with the proper type (12) 
PASS Request interface: new Request("") must inherit property "bodyUsed" with the proper type (13) 
PASS Request interface: new Request("") must inherit property "arrayBuffer" with the proper type (14) 
PASS Request interface: new Request("") must inherit property "blob" with the proper type (15) 
PASS Request interface: new Request("") must inherit property "formData" with the proper type (16) 
PASS Request interface: new Request("") must inherit property "json" with the proper type (17) 
PASS Request interface: new Request("") must inherit property "text" with the proper type (18) 
